agent: Create initial agent structure
This first commit creates the agent overall functioning. This code
initializes the connection through the virtual machine. This is
connected either as a serial port, or as a virtual socket. Scanning
the system allows to determine which type of connection should be
used.

In case of vsock, the connection is pretty simple since we get the
gRPC server listening directly onto this.

In case of serial connection, we introduce an extra layer using yamux
in order to unmux the data coming from the proxy. After the data are
decapsulated from yamux encapsulation, it is passed to the gRPC server
listening on an internal socket instead of vsock.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>